refactor: moved vue-sfc fix to seperate file
